+++ pywal/backends/colorthief.py
@@ -40,8 +40,15 @@ def gen_colors(img):
     return [util.rgb_to_hex(color) for color in raw_colors]
 
 
-def adjust(cols, light, cols16):
-    """Create palette."""
+def adjust(cols, light, **kwargs):
+    """Create palette.
+    :keyword-args:
+    -    c16: use 16 colors through specified method - [ "lighten" | "darken" ]
+    """
+    if "c16" in kwargs:
+        cols16 = kwargs["c16"]
+    else:
+        cols16 = False
     cols.sort(key=util.rgb_to_yiq)
     logging.info("rearranging palette to match ansi colors")
     raw_colors = match.rearrange_palette(cols)
@@ -49,17 +56,17 @@ def adjust(cols, light, cols16):
     raw_colors = [*raw_colors, *raw_colors]
     # print('raw', raw_colors)
 
-    # raw_colors = [*cols, *cols]
-    # for color in raw_colors:
-    #     color = util.lighten_color(color, 0.40)
-    # raw_colors[0] = util.darken_color(cols[0], 0.80)
-
-    adjusted = colors.generic_adjust(raw_colors, light, cols16)
-    # print('adjusted', adjusted)
-    return adjusted
+    return colors.generic_adjust(raw_colors, light, c16=cols16)
 
 
-def get(img, light=False, cols16=False):
-    """Get colorscheme."""
+def get(img, light=False, **kwargs):
+    """Get colorscheme.
+    :keyword-args:
+    -    c16: use 16 colors through specified method - [ "lighten" | "darken" ]
+    """
+    if "c16" in kwargs:
+        cols16 = kwargs["c16"]
+    else:
+        cols16 = False
     cols = gen_colors(img)
-    return adjust(cols, light, cols16)
+    return adjust(cols, light, c16=cols16)